Move sync tests to p1
