+++ hosted/gke/p1/p1_provisioning_test.go
@@ -17,6 +17,13 @@ import (
 var _ = Describe("P1Provisioning", func() {
 	var cluster *management.Cluster
 
+	var _ = BeforeEach(func() {
+		var err error
+		k8sVersion, err = helper.GetK8sVersion(ctx.RancherAdminClient, project, ctx.CloudCred.ID, zone, "", false)
+		Expect(err).To(BeNil())
+		GinkgoLogr.Info(fmt.Sprintf("Using kubernetes version %s for cluster %s", k8sVersion, clusterName))
+	})
+
 	Context("Provisioning a cluster with invalid config", func() {
 
 		AfterEach(func() {
