+++ hosted/gke/p1/sync_provisioning_test.go
@@ -1,4 +1,4 @@
-package sync_test
+package p1_test
 
 import (
 	"fmt"
